Menu Started

+++ sketch_breakout/Block.pde
@@ -6,7 +6,7 @@ class Block extends GameObject
     super(widthScreen/2, heightScreen-50, 50, 50);
   }
 
-  Block(int block_x, int block_y, int block_width, int block_height, color colour)
+  Block(float block_x, float block_y, float block_width, float block_height, color colour)
   {
     super(widthScreen/2, heightScreen-50, 50, 50); 
     this.block_Colour = colour; 
@@ -22,15 +22,13 @@ class Block extends GameObject
 
   void render()
   {
-      if (x+w>width)
-  {
-    x=width-w-30;
-  } 
+    if (x+w>width)
+    {
+      x=width-w-30;
+    } 
 
     noStroke();
     fill(block_Colour);
     rect(x, y, w, h);
   }
-
-
 }
